requirements.txt
# Training dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# ONNX export and optimization
onnx>=1.14.0
onnxruntime>=1.15.0

# CLI and utilities
click>=8.1.0
tqdm>=4.65.0
pyyaml>=6.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
